JSON: move JsonObject to client util
